Move Bluez files into a separate directory to be able to better split things
