+++ src/platform/BUILD.gn
@@ -415,8 +415,6 @@ if (chip_device_platform != "none" && chip_device_platform != "external") {
         "Linux/BLEManagerImpl.cpp",
         "Linux/BLEManagerImpl.h",
         "Linux/BlePlatformConfig.h",
-        "Linux/CHIPBluezHelper.cpp",
-        "Linux/CHIPBluezHelper.h",
         "Linux/CHIPDevicePlatformConfig.h",
         "Linux/CHIPDevicePlatformEvent.h",
         "Linux/CHIPLinuxStorage.cpp",
@@ -440,6 +438,8 @@ if (chip_device_platform != "none" && chip_device_platform != "external") {
         "Linux/PosixConfig.h",
         "Linux/SystemPlatformConfig.h",
         "Linux/SystemTimeSupport.cpp",
+        "Linux/bluez/Helper.cpp",
+        "Linux/bluez/Helper.h",
       ]
 
       if (chip_enable_mdns) {
